Added Splash Screen for Loading
Splash screen launches loading page when application starts and initializes ~4300 Chemical objects. Takes about 31 seconds to initialize all objects. When clicking the inject button, it takes about 28 seconds to load all of the UIInjectionRecord objects on the injectStage. However, currently, the injectStage is non-functional because 25,000+ objects are being initialized and running and consuming memory when you factor in their changeListeners and bindings