Fixed ChemicalLoadingTask Bug
ChemicalLoadingTask was populating casToChemical ConcurrentHashMap with Chemical objects that had a molecular weight of 0. This was causing the peak.plot() method to return NaN values which were being added to the lineChart.dataSeries which was causing the dataSeries to stop displaying any data in the UI. This was because the Chemical.Builder failed to catch my error of not calling .molecularWeight(molecularWeight) on the Chemical object as it was being built in the chemicalLoadingTask. This made me realize Builder design pattern fundamentally not suited for this situation. Chemical object truly requires all 9 of its fields to do its function correctly. I made my Builder with 9 optional fields, stupidly, because I thought builder with 9 required fields is just as weird/wonky as a normal constructor with 9 required fields.